change dir for companents
